requirements.txt
Flask
Faker
Flask-SQLAlchemy
Flask-Migrate
click
psycopg2-binary
flask-marshmallow
marshmallow-sqlalchemy
beautifulsoup4
redis
celery
python-dotenv
pdfkit
flask-login
flask-bcrypt
Flask-WTF
Flask-Mail
email_validator
PyJWT
Pillow
